+++ R/add_tooltip_text.R
@@ -28,7 +28,7 @@
 #'             y = Sepal.Length,
 #'             col = Species,
 #'             text = text,
-#'             theme = gg_theme("helvetica", x_grid = TRUE, y_grid = TRUE)) %>%
+#'             theme = gg_theme("helvetica", grid_v = TRUE, grid_h = TRUE)) %>%
 #'     plotly::ggplotly(tooltip = "text")
 add_tooltip_text <- function(data,
                              ...,